ADD currency exchange icon button in sliver
