--- lib/features/home/ui/widgets/currency_exchange_component/sliver_currencies_list_title.dart
@@ -1,20 +1,0 @@
-import 'package:currencypro/core/utils/app_text_style.dart';
-import 'package:flutter/material.dart';
-
-class SliverCurrenciesListTitle extends StatelessWidget {
-  const SliverCurrenciesListTitle({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SliverToBoxAdapter(
-      child: Text(
-        'Currencies',
-        style: AppTextStyle.textStyle19.copyWith(
-          fontWeight: FontWeight.w600,
-        ),
-      ),
-    );
-  }
-}
